Refactoring by changing the method signature
Using the refactoring tools in eclipse, the method signature of the
method playAction() was changed and the String parameter of type was
added. This required some manual refactoring on my part as well and I
simply added the boolean return statement of false to each of the method
calls to playAction() that were affected by the change to the method
signature.